+++ resolvers.ts
@@ -7,12 +7,24 @@ import {
   query_fedWorkflowRunsAggregate_items,
   query_fedWorkflowRuns_items,
 } from "./.mesh";
-import { get, postWithCSRF, shouldReadFromNextGen } from "./utils/httpUtils";
+import {
+  fetchFromNextGen,
+  getFromRails,
+  get,
+  getFromRails,
+  postWithCSRF,
+  shouldReadFromNextGen,
+} from "./utils/httpUtils";
 import {
   formatTaxonHits,
   formatTaxonLineage,
 } from "./utils/mngsWorkflowResultsUtils";
 import { formatUrlParams } from "./utils/paramsUtils";
+import {
+  convertSequencingReadsQuery,
+  convertWorkflowRunsQuery,
+} from "./utils/queryFormatUtils";
+import { isRunFinalized, parseRefFasta } from "./utils/responseHelperUtils";
 
 /**
  * Arbitrary very large number used temporarily during Rails read phase to force Rails not to
@@ -248,119 +260,124 @@ export const resolvers: Resolvers = {
           },
         ];
         return ret;
-      } else {
-        // The comments in the formatUrlParams() call correspond to the line in the current
-        // codebase's callstack where the params are set, so help ensure we're not missing anything.
-        const { workflow_runs } = await get({
-          url:
-            "/workflow_runs.json" +
-            formatUrlParams({
-              // index.ts
-              // const getWorkflowRuns = ({
-              mode: "with_sample_info",
-              //  - DiscoveryDataLayer.ts
-              //    await this._collection.fetchDataCallback({
-              domain: input?.todoRemove?.domain,
-              //  -- DiscoveryView.tsx
-              //     ...this.getConditions(workflow)
-              projectId: input?.todoRemove?.projectId,
-              search: input?.todoRemove?.search,
-              orderBy: input?.todoRemove?.orderBy,
-              orderDir: input?.todoRemove?.orderDir,
-              //  --- DiscoveryView.tsx
-              //      filters: {
-              host: input?.todoRemove?.host,
-              locationV2: input?.todoRemove?.locationV2,
-              taxon: input?.todoRemove?.taxons,
-              taxaLevels: input?.todoRemove?.taxaLevels,
-              time: input?.todoRemove?.time,
-              tissue: input?.todoRemove?.tissue,
-              visibility: input?.todoRemove?.visibility,
-              workflow: input?.todoRemove?.workflow,
-              //  - DiscoveryDataLayer.ts
-              //    await this._collection.fetchDataCallback({
-              limit: input?.limit,
-              offset: input?.offset,
-              listAllIds: false,
-            }),
-          args,
-          context,
-        });
-        if (!workflow_runs?.length) {
-          return [];
-        }
+      }
 
-        return workflow_runs.map((run): query_fedConsensusGenomes_items => {
-          const inputs = run.inputs;
-          const qualityMetrics = run.cached_results?.quality_metrics;
-          const sample = run.sample;
-          const sampleInfo = sample?.info;
-          const sampleMetadata = sample?.metadata;
+      // DISCOVERY VIEW:
+      // The comments in the formatUrlParams() call correspond to the line in the current
+      // codebase's callstack where the params are set, so help ensure we're not missing anything.
+      const { workflow_runs } = await get({
+        url:
+          "/workflow_runs.json" +
+          formatUrlParams({
+            // index.ts
+            // const getWorkflowRuns = ({
+            mode: "with_sample_info",
+            //  - DiscoveryDataLayer.ts
+            //    await this._collection.fetchDataCallback({
+            domain: input?.todoRemove?.domain,
+            //  -- DiscoveryView.tsx
+            //     ...this.getConditions(workflow)
+            projectId: input?.todoRemove?.projectId,
+            search: input?.todoRemove?.search,
+            orderBy: input?.todoRemove?.orderBy,
+            orderDir: input?.todoRemove?.orderDir,
+            //  --- DiscoveryView.tsx
+            //      filters: {
+            host: input?.todoRemove?.host,
+            locationV2: input?.todoRemove?.locationV2,
+            taxon: input?.todoRemove?.taxons,
+            taxaLevels: input?.todoRemove?.taxaLevels,
+            time: input?.todoRemove?.time,
+            tissue: input?.todoRemove?.tissue,
+            visibility: input?.todoRemove?.visibility,
+            workflow: input?.todoRemove?.workflow,
+            //  - DiscoveryDataLayer.ts
+            //    await this._collection.fetchDataCallback({
+            limit: input?.limit,
+            offset: input?.offset,
+            listAllIds: false,
+          }),
+        args,
+        context,
+      });
+      if (!workflow_runs?.length) {
+        return [];
+      }
 
-          const taxon =
-            inputs?.taxon_name != null
-              ? {
-                  name: inputs.taxon_name,
-                }
-              : null;
-          return {
-            producingRunId: run.id?.toString(),
+      return workflow_runs.map((run): query_fedConsensusGenomes_items => {
+        const inputs = run.inputs;
+        const qualityMetrics = run.cached_results?.quality_metrics;
+        const sample = run.sample;
+        const sampleInfo = sample?.info;
+        const sampleMetadata = sample?.metadata;
+
+        const taxon =
+          inputs?.taxon_name != null
+            ? {
+                name: inputs.taxon_name,
+              }
+            : null;
+        const accession =
+          inputs?.accession_id != null && inputs?.accession_name != null
+            ? {
+                accessionId: inputs?.accession_id,
+                accessionName: inputs?.accession_name,
+              }
+            : null;
+        return {
+          producingRunId: run.id?.toString(),
+          taxon,
+          accession,
+          metrics: {
+            coverageDepth: run.cached_results?.coverage_viz?.coverage_depth,
+            totalReads: qualityMetrics?.total_reads,
+            gcPercent: qualityMetrics?.gc_percent,
+            refSnps: qualityMetrics?.ref_snps,
+            percentIdentity: qualityMetrics?.percent_identity,
+            nActg: qualityMetrics?.n_actg,
+            percentGenomeCalled: qualityMetrics?.percent_genome_called,
+            nMissing: qualityMetrics?.n_missing,
+            nAmbiguous: qualityMetrics?.n_ambiguous,
+            referenceGenomeLength: qualityMetrics?.reference_genome_length,
+          },
+          sequencingRead: {
+            nucleicAcid: sampleMetadata?.nucleotide_type ?? "",
+            protocol: inputs?.wetlab_protocol,
+            medakaModel: inputs?.medaka_model,
+            technology: inputs?.technology ?? "",
             taxon,
-            referenceGenome: {
-              accessionId: inputs?.accession_id,
-              accessionName: inputs?.accession_name,
-            },
-            metrics: {
-              coverageDepth: run.cached_results?.coverage_viz?.coverage_depth,
-              totalReads: qualityMetrics?.total_reads,
-              gcPercent: qualityMetrics?.gc_percent,
-              refSnps: qualityMetrics?.ref_snps,
-              percentIdentity: qualityMetrics?.percent_identity,
-              nActg: qualityMetrics?.n_actg,
-              percentGenomeCalled: qualityMetrics?.percent_genome_called,
-              nMissing: qualityMetrics?.n_missing,
-              nAmbiguous: qualityMetrics?.n_ambiguous,
-              referenceGenomeLength: qualityMetrics?.reference_genome_length,
-            },
-            sequencingRead: {
-              nucleicAcid: sampleMetadata?.nucleotide_type ?? "",
-              protocol: inputs?.wetlab_protocol,
-              medakaModel: inputs?.medaka_model,
-              technology: inputs?.technology ?? "",
-              taxon,
-              sample: {
-                railsSampleId: sample?.id,
-                name: sampleInfo?.name ?? "",
-                notes: sampleInfo?.sample_notes,
-                uploadError: sampleInfo?.result_status_description,
-                collectionLocation:
-                  typeof sampleMetadata?.collection_location_v2 === "string"
-                    ? sampleMetadata.collection_location_v2
-                    : sampleMetadata?.collection_location_v2?.name ?? "",
-                sampleType: sampleMetadata?.sample_type ?? "",
-                waterControl: sampleMetadata?.water_control === "Yes",
-                hostOrganism:
-                  sampleInfo?.host_genome_name != null
-                    ? {
-                        name: sampleInfo.host_genome_name,
-                      }
-                    : null,
-                collection: {
-                  name: sample?.project_name,
-                  public: Boolean(sampleInfo?.public),
-                },
-                ownerUserId: sample?.uploader?.id,
-                // TODO: Make runner come from Workflows stitched with the user service when NextGen
-                // ready.
-                ownerUserName: run.runner?.name ?? sample?.uploader?.name,
-                metadatas: {
-                  edges: getMetadataEdges(sampleMetadata),
-                },
+            sample: {
+              railsSampleId: sample?.id,
+              name: sampleInfo?.name ?? "",
+              notes: sampleInfo?.sample_notes,
+              uploadError: sampleInfo?.result_status_description,
+              collectionLocation:
+                typeof sampleMetadata?.collection_location_v2 === "string"
+                  ? sampleMetadata.collection_location_v2
+                  : sampleMetadata?.collection_location_v2?.name ?? "",
+              sampleType: sampleMetadata?.sample_type ?? "",
+              waterControl: sampleMetadata?.water_control === "Yes",
+              hostOrganism:
+                sampleInfo?.host_genome_name != null
+                  ? {
+                      name: sampleInfo.host_genome_name,
+                    }
+                  : null,
+              collection: {
+                name: sample?.project_name,
+                public: Boolean(sampleInfo?.public),
+              },
+              ownerUserId: sample?.uploader?.id,
+              // TODO: Make runner come from Workflows stitched with the user service when NextGen
+              // ready.
+              ownerUserName: run.runner?.name ?? sample?.uploader?.name,
+              metadatas: {
+                edges: getMetadataEdges(sampleMetadata),
               },
             },
-          };
-        });
-      }
+          },
+        };
+      });
     },
     ConsensusGenomeWorkflowResults: async (root, args, context, info) => {
       const { coverage_viz, quality_metrics, taxon_info } = await get({
@@ -455,6 +472,218 @@ export const resolvers: Resolvers = {
       } catch {
         return res;
       }
+    },
+    SampleForReport: async (root, args, context) => {
+      /* --------------------- Rails and Next Gen --------------------- */
+      const sampleInfo = await getFromRails({
+        url: `/samples/${args.railsSampleId}.json`,
+        args,
+        context,
+      });
+      console.log("sampleInfo - getFromRails", JSON.stringify(sampleInfo));
+      // Make output acceptable to Relay - convert ids to strings
+      if (sampleInfo?.pipeline_runs) {
+        const updatedPipelineRuns = sampleInfo?.pipeline_runs.map(
+          pipelineRun => {
+            return {
+              ...pipelineRun,
+              id: pipelineRun.id.toString(),
+            };
+          },
+        );
+        sampleInfo.pipeline_runs = updatedPipelineRuns;
+      }
+      if (sampleInfo?.workflow_runs) {
+        const updatedWorkflowRuns = sampleInfo?.workflow_runs.map(
+          workflowRun => {
+            return {
+              ...workflowRun,
+              id: workflowRun.id.toString(),
+            };
+          },
+        );
+        sampleInfo.workflow_runs = updatedWorkflowRuns;
+      }
+      if (sampleInfo?.project) {
+        sampleInfo.project.id = sampleInfo.project.id.toString();
+      }
+
+      const nextGenEnabled = await shouldReadFromNextGen(context);
+      /* --------------------- Rails --------------------- */
+      if (!nextGenEnabled) {
+        console.log("next gen off Response", {
+          id: `sample-for-query-${args.railsSampleId}`,
+          railsSampleId: args.railsSampleId,
+          ...sampleInfo,
+        });
+        return {
+          id: `sample-for-query-${args.railsSampleId}`,
+          railsSampleId: args.railsSampleId,
+          ...sampleInfo,
+        };
+      }
+      /* --------------------- Next Gen --------------------- */
+
+      // NextGen Steps:
+      // continue using everything from rails except for workflow_runs
+      // get sample from rails using railsSampleId (same as above). This includes:
+      // -- sample info
+      // -- pipeline runs
+      // -- workflowRuns:
+      // ---- AMR workflow runs
+      // ---- pre-migration CG workflow runs
+      // ---- dual write CG workflow runs
+      // query entities using railsSampleId to get NextGenSampleId and done CG workflow runs
+      // query workflows using NextGenSampleId to get in progress CG workflow runs
+      // combine workflow data from entities and workflows
+      // deduplicate between rails and next gen
+
+      const entitiesQuery = `
+          query EntitiesQuery {
+            samples(where: {railsSampleId: {_eq: ${args.railsSampleId}}}) {
+              id
+              sequencingReads {
+                edges {
+                  node {
+                    consensusGenomes {
+                      edges {
+                        node {
+                          id
+                          createdAt
+                          producingRunId
+                          referenceGenome {
+                            id
+                            file {
+                              path
+                            }
+                          }
+                          accession {
+                            accessionId
+                            accessionName
+                          }
+                          taxon {
+                            id
+                            name
+                          }
+                          sequencingRead {
+                            technology
+                          }
+                        }
+                      }
+                    }
+                    sample {
+                      hostOrganism {
+                        id
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          } 
+        `;
+      const entitiesResp = await get({
+        args,
+        context,
+        serviceType: "entities",
+        customQuery: entitiesQuery,
+      });
+      console.log("entitiesResp - get 1", JSON.stringify(entitiesResp));
+
+      // Query workflows using NextGenSampleId to get in progress CG workflow runs
+      const nextGenSampleId = entitiesResp?.data.samples[0].id;
+      const workflowsQuery = `
+          query WorkflowsQuery {
+            workflowRuns(where: {entityInputs: {inputEntityId: {_eq: "${nextGenSampleId}"}}}) {
+              id
+              _id
+              railsWorkflowRunId
+              status
+              ownerUserId
+              errorMessage
+              workflowVersion {
+                version
+                id
+                workflow {
+                  name
+                }
+              }
+              createdAt
+              endedAt
+              rawInputsJson
+            }
+          }
+      `;
+      const workflowsResp = await get({
+        args,
+        context,
+        serviceType: "workflows",
+        customQuery: workflowsQuery,
+      });
+      console.log("workflowsResp - get 2", JSON.stringify(workflowsResp));
+      const consensusGenomes =
+        entitiesResp.data.samples[0].sequencingReads.edges[0].node
+          .consensusGenomes.edges;
+      const workflowsWorkflowRuns = workflowsResp?.data?.workflowRuns || [];
+      const nextGenWorkflowRuns = workflowsWorkflowRuns.map(workflowRun => {
+        const consensusGenome = consensusGenomes.find(consensusGenome => {
+          return consensusGenome.node.producingRunId === workflowRun.id;
+        });
+        const { accession, taxon, sequencingRead } =
+          consensusGenome?.node || {};
+        const parsedRawInputsJson = JSON.parse(workflowRun.rawInputsJson);
+        // If !consensusGenome this is a workflow run that is in progress
+        return {
+          deprecated: workflowRun?.deprecated_by,
+          executed_at: workflowRun?.createdAt,
+          id: workflowRun?.id,
+          input_error: workflowRun?.errorMessage,
+          inputs: {
+            accession_id: accession?.accessionId,
+            accession_name: accession?.accessionName,
+            creation_source: parsedRawInputsJson?.creation_source,
+            ref_fasta: parseRefFasta(
+              consensusGenome?.node?.referenceGenome?.file?.path,
+            ),
+            taxon_id: taxon?.id,
+            taxon_name: taxon?.name,
+            technology: sequencingRead?.technology,
+          },
+          rails_workflow_run_id: workflowRun?.railsWorkflowRunId, // this is added for deduplicating below
+          run_finalized: isRunFinalized(workflowRun?.status),
+          status: workflowRun?.status,
+          wdl_version: workflowRun?.workflowVersion.version,
+          workflow: workflowRun?.workflowVersion.workflow.name,
+        };
+      });
+      console.log("nextGenWorkflowRuns", nextGenWorkflowRuns);
+      // Deduplicate sampleInfo.workflow_runs(from Rails) and nextGenWorkflowRuns(from NextGen)
+      let dedupedWorkflowRuns;
+      dedupedWorkflowRuns = [...nextGenWorkflowRuns];
+      console.log("sampleInfo.workflow_runs", sampleInfo.workflow_runs);
+      for (const railsWorkflowRun of sampleInfo.workflow_runs) {
+        const alreadyExists = nextGenWorkflowRuns.find(
+          nextGenWorkflowRun =>
+            nextGenWorkflowRun.rails_workflow_run_id.toString() ===
+            railsWorkflowRun.id,
+        );
+        if (!alreadyExists) {
+          dedupedWorkflowRuns.push(railsWorkflowRun);
+        }
+      }
+      console.log("dedupedWorkflowRuns", dedupedWorkflowRuns);
+      console.log("return next gen enabled", {
+        id: args.railsSampleId,
+        railsSampleId: args.railsSampleId,
+        ...sampleInfo,
+        workflow_runs: dedupedWorkflowRuns,
+      });
+      return {
+        id: args.railsSampleId,
+        railsSampleId: args.railsSampleId,
+        ...sampleInfo,
+        workflow_runs: dedupedWorkflowRuns,
+      };
     },
     MngsWorkflowResults: async (root, args, context, info) => {
       const data = await get({
@@ -587,8 +816,71 @@ export const resolvers: Resolvers = {
         };
       });
     },
-    fedSequencingReads: async (root, args, context) => {
+    fedSequencingReads: async (root, args, context: any) => {
       const input = args.input;
+
+      // NEXT GEN:
+      const nextGenEnabled = await shouldReadFromNextGen(context);
+      if (nextGenEnabled) {
+        const nextGenSequencingReads = (
+          await fetchFromNextGen({
+            customQuery: convertSequencingReadsQuery(context.params.query),
+            customVariables: {
+              where: input.where,
+              orderBy: input.orderBy != null ? [input.orderBy] : [], // TODO: Migrate to array orderBy.
+              limitOffset: input.limitOffset,
+              producingRunIds: input?.where?.id?._in,
+            },
+            serviceType: "entities",
+            args,
+            context,
+          })
+        ).data.sequencingReads;
+        const railsSamples = (
+          await getFromRails({
+            url:
+              "/samples/index_v2.json" +
+              formatUrlParams({
+                sampleIds: nextGenSequencingReads.map(
+                  sequencingRead => sequencingRead.sample.railsSampleId,
+                ),
+                limit: TEN_MILLION,
+                offset: 0,
+                listAllIds: false,
+              }),
+            args,
+            context,
+          })
+        ).samples;
+
+        const samplesById = new Map<number, any>(
+          railsSamples.map(sample => [sample.id, sample]),
+        );
+        for (const nextGenSequencingRead of nextGenSequencingReads) {
+          const nextGenSample = nextGenSequencingRead.sample;
+          const railsSample = samplesById.get(nextGenSample.railsSampleId);
+          const railsMetadata = railsSample.details?.metadata;
+          const railsDbSample = railsSample.details?.db_sample;
+
+          nextGenSequencingRead.nucleicAcid =
+            railsMetadata?.nucleotide_type ?? "";
+          nextGenSample.collectionLocation =
+            railsMetadata?.collection_location_v2 ?? "";
+          nextGenSample.sampleType = railsMetadata?.sample_type ?? "";
+          nextGenSample.waterControl = railsMetadata?.water_control === "Yes";
+          nextGenSample.notes = railsDbSample?.sample_notes;
+          nextGenSample.uploadError = railsDbSample?.upload_error;
+          nextGenSample.collection = {
+            name: railsSample.details?.derived_sample_output?.project_name,
+            public: railsSample.public === 1,
+          };
+          nextGenSample.metadatas = {
+            edges: getMetadataEdges(railsMetadata),
+          };
+        }
+
+        return nextGenSequencingReads;
+      }
 
       // The comments in the formatUrlParams() call correspond to the line in the current
       // codebase's callstack where the params are set, so help ensure we're not missing anything.
@@ -620,8 +912,8 @@ export const resolvers: Resolvers = {
             workflow: input?.todoRemove?.workflow,
             //  - DiscoveryDataLayer.ts
             //    await this._collection.fetchDataCallback({
-            limit: input?.limit,
-            offset: input?.offset,
+            limit: input?.limit ?? input?.limitOffset?.limit, // TODO: Just use limitOffset.
+            offset: input?.offset ?? input?.limitOffset?.offset,
             listAllIds: false,
           }),
         args,
@@ -647,14 +939,19 @@ export const resolvers: Resolvers = {
                 name: inputs.taxon_name,
               }
             : null;
+        const accession =
+          inputs?.accession_id != null && inputs?.accession_name != null
+            ? {
+                accessionId: inputs?.accession_id,
+                accessionName: inputs?.accession_name,
+              }
+            : null;
         const consensusGenomeEdge = {
           node: {
             producingRunId: run.id?.toString(),
             taxon,
-            referenceGenome: {
-              accessionId: inputs?.accession_id,
-              accessionName: inputs?.accession_name,
-            },
+            referenceGenome: accession,
+            accession,
             metrics: {
               coverageDepth: run.cached_results?.coverage_viz?.coverage_depth,
               totalReads: qualityMetrics?.total_reads,
@@ -813,9 +1110,10 @@ export const resolvers: Resolvers = {
       });
       return annotations;
     },
-    fedWorkflowRuns: async (root, args, context) => {
+    fedWorkflowRuns: async (root, args, context: any) => {
       const input = args.input;
 
+      // CG REPORT:
       // If we provide a list of workflowRunIds, we assume that this is for getting valid consensus genome workflow runs.
       // This endpoint only provides id, ownerUserId, and status.
       if (input?.where?.id?._in && typeof input?.where?.id?._in === "object") {
@@ -836,6 +1134,22 @@ export const resolvers: Resolvers = {
         }));
       }
 
+      // DISCOVERY VIEW:
+      const nextGenEnabled = await shouldReadFromNextGen(context);
+      if (nextGenEnabled) {
+        const response = await fetchFromNextGen({
+          customQuery: convertWorkflowRunsQuery(context.params.query),
+          customVariables: {
+            where: input.where,
+            orderBy: input.orderBy != null ? [input.orderBy] : [], // TODO: Migrate to array orderBy.
+          },
+          serviceType: "workflows",
+          args,
+          context,
+        });
+        return response.data.workflowRuns;
+      }
+
       // TODO(bchu): Remove all the non-Workflows fields after moving and integrating them into the
       // Entities call.
       const { workflow_runs } = await get({
@@ -846,8 +1160,6 @@ export const resolvers: Resolvers = {
             domain: input?.todoRemove?.domain,
             projectId: input?.todoRemove?.projectId,
             search: input?.todoRemove?.search,
-            // TODO: Cover sorting by time, version, or creation source (though Rails doesn't
-            // actually support the latter 2).
             orderBy: input?.todoRemove?.orderBy,
             orderDir: input?.todoRemove?.orderDir,
             host: input?.todoRemove?.host,
@@ -875,10 +1187,11 @@ export const resolvers: Resolvers = {
           ownerUserId: run.runner?.id?.toString(),
           startedAt: run.created_at,
           status: run.status,
+          rawInputsJson: `{"creation_source": "${run.inputs?.creation_source ?? ""}"}`,
           workflowVersion: {
             version: run.wdl_version,
             workflow: {
-              name: run.inputs?.creation_source,
+              name: run.inputs?.creation_source, // TODO: Delete when FE uses rawInputsJson.
             },
           },
           entityInputs: {