+++ tests/SequencingReadsQuery.test.ts
@@ -2,11 +2,19 @@ import { ExecuteMeshFn } from "@graphql-mesh/runtime";
 import { getMeshInstance } from "./utils/MeshInstance";
 import * as httpUtils from "../utils/httpUtils";
 import { getExampleQuery } from "./utils/ExampleQueryFiles";
+import { assertEqualsNoWhitespace } from "./utils/StringUtils";
+import { convertSequencingReadsQuery } from "../utils/queryFormatUtils";
 
 jest.spyOn(httpUtils, "get");
+jest.spyOn(httpUtils, "shouldReadFromNextGen");
+jest.spyOn(httpUtils, "fetchFromNextGen");
+jest.spyOn(httpUtils, "getFromRails");
 
 beforeEach(() => {
   (httpUtils.get as jest.Mock).mockClear();
+  (httpUtils.shouldReadFromNextGen as jest.Mock).mockClear();
+  (httpUtils.fetchFromNextGen as jest.Mock).mockClear();
+  (httpUtils.getFromRails as jest.Mock).mockClear();
 });
 
 const query = getExampleQuery("sequencing-reads-query");
@@ -15,6 +23,9 @@ describe("sequencingReads query:", () => {
   let execute: ExecuteMeshFn;
 
   beforeEach(async () => {
+    (httpUtils.shouldReadFromNextGen as jest.Mock).mockImplementation(() =>
+      Promise.resolve(false),
+    );
     const mesh$ = await getMeshInstance();
     ({ execute } = mesh$);
   });
@@ -120,7 +131,10 @@ describe("sequencingReads query:", () => {
 
     const result = await execute(query, {});
 
-    const metadataFields = result.data.fedSequencingReads[0].sample.metadatas.edges.map(edge => edge.node.fieldName);
+    const metadataFields =
+      result.data.fedSequencingReads[0].sample.metadatas.edges.map(
+        edge => edge.node.fieldName,
+      );
     expect(metadataFields).toHaveLength(3);
     expect(metadataFields[0]).toEqual("key1");
     expect(metadataFields[1]).toEqual("key2");
@@ -142,7 +156,9 @@ describe("sequencingReads query:", () => {
 
     const result = await execute(query, {});
 
-    expect(result.data.fedSequencingReads[0].sample.metadatas.edges).toHaveLength(0);
+    expect(
+      result.data.fedSequencingReads[0].sample.metadatas.edges,
+    ).toHaveLength(0);
   });
 
   it("Only returns taxon object if name exists", async () => {
@@ -166,7 +182,9 @@ describe("sequencingReads query:", () => {
 
     const result = await execute(query, {});
 
-    expect(result.data.fedSequencingReads[0].sample.collectionLocation).toBe("");
+    expect(result.data.fedSequencingReads[0].sample.collectionLocation).toBe(
+      "",
+    );
   });
 
   it("Returns string location field", async () => {
@@ -184,7 +202,9 @@ describe("sequencingReads query:", () => {
 
     const result = await execute(query, {});
 
-    expect(result.data.fedSequencingReads[0].sample.collectionLocation).toBe("Redwood City");
+    expect(result.data.fedSequencingReads[0].sample.collectionLocation).toBe(
+      "Redwood City",
+    );
   });
 
   it("Returns object name for location field", async () => {
@@ -204,7 +224,9 @@ describe("sequencingReads query:", () => {
 
     const result = await execute(query, {});
 
-    expect(result.data.fedSequencingReads[0].sample.collectionLocation).toBe("Redwood City");
+    expect(result.data.fedSequencingReads[0].sample.collectionLocation).toBe(
+      "Redwood City",
+    );
   });
 
   it("Converts water control to boolean", async () => {
@@ -281,16 +303,320 @@ describe("sequencingReads query:", () => {
 
     expect(sequencingReads[0].id).toBe("123");
     expect(sequencingReads[0].consensusGenomes.edges.length).toBe(3);
-    expect(sequencingReads[0].consensusGenomes.edges[0].node.producingRunId).toBe("a");
-    expect(sequencingReads[0].consensusGenomes.edges[0].node.taxon.name).toBe("Taxon1");
-    expect(sequencingReads[0].consensusGenomes.edges[1].node.producingRunId).toBe("b");
-    expect(sequencingReads[0].consensusGenomes.edges[1].node.taxon.name).toBe("Taxon2");
-    expect(sequencingReads[0].consensusGenomes.edges[2].node.producingRunId).toBe("d");
-    expect(sequencingReads[0].consensusGenomes.edges[2].node.taxon.name).toBe("Taxon1");
+    expect(
+      sequencingReads[0].consensusGenomes.edges[0].node.producingRunId,
+    ).toBe("a");
+    expect(sequencingReads[0].consensusGenomes.edges[0].node.taxon.name).toBe(
+      "Taxon1",
+    );
+    expect(
+      sequencingReads[0].consensusGenomes.edges[1].node.producingRunId,
+    ).toBe("b");
+    expect(sequencingReads[0].consensusGenomes.edges[1].node.taxon.name).toBe(
+      "Taxon2",
+    );
+    expect(
+      sequencingReads[0].consensusGenomes.edges[2].node.producingRunId,
+    ).toBe("d");
+    expect(sequencingReads[0].consensusGenomes.edges[2].node.taxon.name).toBe(
+      "Taxon1",
+    );
 
     expect(sequencingReads[1].id).toBe("456");
     expect(sequencingReads[1].consensusGenomes.edges.length).toBe(1);
-    expect(sequencingReads[1].consensusGenomes.edges[0].node.producingRunId).toBe("c");
-    expect(sequencingReads[1].consensusGenomes.edges[0].node.taxon.name).toBe("Taxon3");
+    expect(
+      sequencingReads[1].consensusGenomes.edges[0].node.producingRunId,
+    ).toBe("c");
+    expect(sequencingReads[1].consensusGenomes.edges[0].node.taxon.name).toBe(
+      "Taxon3",
+    );
+  });
+
+  it("Constructs correct NextGen query", () => {
+    const query = `
+      query DiscoveryViewFCSequencingReadsQuery(
+        $input: queryInput_fedSequencingReads_input_Input
+      ) {
+        fedSequencingReads(input: $input) {
+          id
+          nucleicAcid
+          protocol
+          medakaModel
+          technology
+          taxon {
+            name
+          }
+          sample {
+            railsSampleId
+            name
+            notes
+            collectionLocation
+            sampleType
+            waterControl
+            uploadError
+            hostOrganism {
+              name
+            }
+            collection {
+              name
+              public
+            }
+            ownerUserId
+            ownerUserName
+            metadatas {
+              edges {
+                node {
+                  fieldName
+                  value
+                }
+              }
+            }
+          }
+          consensusGenomes {
+            edges {
+              node {
+                producingRunId
+                taxon {
+                  name
+                }
+                accession {
+                  accessionId
+                  accessionName
+                }
+                metrics {
+                  coverageDepth
+                  totalReads
+                  gcPercent
+                  refSnps
+                  percentIdentity
+                  nActg
+                  percentGenomeCalled
+                  nMissing
+                  nAmbiguous
+                  referenceGenomeLength
+                }
+              }
+            }
+          }
+        }
+      }`;
+
+    assertEqualsNoWhitespace(
+      convertSequencingReadsQuery(query),
+      `query ($where: SequencingReadWhereClause,
+              $orderBy: [SequencingReadOrderByClause!],
+              $limitOffset: LimitOffsetClause,
+              $producingRunIds: [UUID!]) {
+        sequencingReads(where: $where, orderBy: $orderBy, limitOffset: $limitOffset) {
+          id
+          protocol
+          medakaModel
+          technology
+          taxon {
+            name
+          }
+          sample {
+            railsSampleId
+            name
+            hostOrganism {
+              name
+            }
+            ownerUserId
+            metadatas {
+              edges {
+                node {
+                  fieldName
+                  value
+                }
+              }
+            }
+          }
+          consensusGenomes(where: { producingRunId: { _in: $producingRunIds } }) {
+            edges {
+              node {
+                producingRunId
+                taxon {
+                  name
+                }
+                accession {
+                  accessionId
+                  accessionName
+                }
+                metrics {
+                  coverageDepth
+                  totalReads
+                  gcPercent
+                  refSnps
+                  percentIdentity
+                  nActg
+                  percentGenomeCalled
+                  nMissing
+                  nAmbiguous
+                  referenceGenomeLength
+                }
+              }
+            }
+          }
+        }
+      }`,
+    );
+  });
+
+  it("Joins NextGen and Rails data", async () => {
+    (httpUtils.shouldReadFromNextGen as jest.Mock).mockImplementation(() =>
+      Promise.resolve(true),
+    );
+    (httpUtils.fetchFromNextGen as jest.Mock).mockImplementation(() =>
+      Promise.resolve({
+        data: {
+          sequencingReads: [
+            {
+              id: "abc",
+              sample: {
+                railsSampleId: 123,
+              },
+              technology: "Technology A",
+              consensusGenomes: {
+                edges: [],
+              },
+            },
+            {
+              id: "def",
+              sample: {
+                railsSampleId: 123,
+              },
+              technology: "Technology B",
+              consensusGenomes: {
+                edges: [],
+              },
+            },
+            {
+              id: "ghi",
+              sample: {
+                railsSampleId: 456,
+              },
+              technology: "Technology C",
+              consensusGenomes: {
+                edges: [],
+              },
+            },
+          ],
+        },
+      }),
+    );
+    (httpUtils.getFromRails as jest.Mock).mockImplementation(() =>
+      Promise.resolve({
+        samples: [
+          {
+            id: 123,
+            details: {
+              db_sample: {
+                sample_notes: "Note A",
+              },
+              derived_sample_output: {
+                project_name: "Project A",
+              },
+              metadata: {
+                collection_location_v2: "USA",
+                custom1: "Custom value 1",
+              },
+            },
+          },
+          {
+            id: 456,
+            details: {
+              db_sample: {
+                sample_notes: "Note B",
+              },
+              derived_sample_output: {
+                project_name: "Project B",
+              },
+              metadata: {
+                collection_location_v2: "Mexico",
+                custom2: "Custom value 2",
+              },
+            },
+          },
+        ],
+      }),
+    );
+
+    const sequencingReads = (await execute(query, {}, { params: { query } }))
+      .data.fedSequencingReads;
+
+    expect(sequencingReads).toMatchObject([
+      expect.objectContaining({
+        id: "abc",
+        technology: "Technology A",
+        sample: {
+          railsSampleId: 123,
+          collectionLocation: "USA",
+          notes: "Note A",
+          collection: {
+            name: "Project A",
+            public: false,
+          },
+          metadatas: {
+            edges: [
+              {
+                node: {
+                  fieldName: "custom1",
+                  value: "Custom value 1",
+                },
+              },
+            ],
+          },
+          waterControl: false,
+        },
+      }),
+      expect.objectContaining({
+        id: "def",
+        technology: "Technology B",
+        sample: {
+          railsSampleId: 123,
+          collectionLocation: "USA",
+          notes: "Note A",
+          collection: {
+            name: "Project A",
+            public: false,
+          },
+          metadatas: {
+            edges: [
+              {
+                node: {
+                  fieldName: "custom1",
+                  value: "Custom value 1",
+                },
+              },
+            ],
+          },
+          waterControl: false,
+        },
+      }),
+      expect.objectContaining({
+        id: "ghi",
+        technology: "Technology C",
+        sample: {
+          railsSampleId: 456,
+          collectionLocation: "Mexico",
+          notes: "Note B",
+          collection: {
+            name: "Project B",
+            public: false,
+          },
+          metadatas: {
+            edges: [
+              {
+                node: {
+                  fieldName: "custom2",
+                  value: "Custom value 2",
+                },
+              },
+            ],
+          },
+          waterControl: false,
+        },
+      }),
+    ]);
   });
 });